+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Crop-circles".
 - Tests for "Nicholas Breakspeare and Hugh of Evesham".
 - Tests for "Nature of triangles".
+- Tests for "Artificial emotional intelligence".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".